requirements.txt
# requirements.txt
openai
pyyaml
termcolor
keyboard
pyaudio
Pillow
pyautogui
pygame
opencv-python
moviepy
nltk
pygments